requirements.txt
fastapi>=0.85
uvicorn[standard]>=0.18
pyshark>=0.4.3
streamlit>=1.25
pandas>=1.5
requests>=2.28
python-dotenv>=0.21
plotly
netifaces